fix(github): extract action-preset service to keep service.go under line limit
Splits the preset CRUD methods out of the oversized service.go and applies prettier formatting to three frontend files flagged by CI.
